restructure fe
