+++ frontend/src/index.css
@@ -1,62 +1,27 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
 
 @layer base {
-  [type="text"],
-  [type="email"],
-  [type="url"],
-  [type="password"],
-  [type="number"],
-  [type="date"],
-  [type="datetime-local"],
-  [type="month"],
-  [type="search"],
-  [type="tel"],
-  [type="time"],
-  [type="week"],
-  [multiple],
-  textarea,
-  select {
-    @apply w-full border-gray-300 rounded-md shadow-sm focus:border-black focus:ring-black my-3;
+    [type="text"],
+    [type="email"],
+    [type="url"],
+    [type="password"],
+    [type="number"],
+    [type="date"],
+    [type="datetime-local"],
+    [type="month"],
+    [type="search"],
+    [type="tel"],
+    [type="time"],
+    [type="week"],
+    [multiple],
+    textarea,
+    select {
+      @apply w-full border-gray-300 rounded-md shadow-sm focus:border-black focus:ring-black my-3;
+    }
+    [type="checkbox"],
+    [type="radio"] {
+      @apply border-gray-300 rounded-md focus:border-black focus:ring-black my-3;
+    }
   }
-  [type="checkbox"],
-  [type="radio"] {
-    @apply border-gray-300 rounded-md focus:border-black focus:ring-black my-3;
-  }
-}
-      
-
-@layer{
-  .link-underline {
-    border-bottom-width: 0;
-    background-image: linear-gradient(transparent, transparent), linear-gradient(#fff, #fff);
-    background-size: 0 3px;
-    background-position: 0 100%;
-    background-repeat: no-repeat;
-    transition: background-size .3s ease-in-out;
-}
-
-.link-underline-black {
-    background-image: linear-gradient(transparent, transparent), linear-gradient(#000, #000)
-}
-
-.link-underline:hover {
-    background-size: 100% 3px;
-    background-position: 0 100%
-}
-}
