--- frontend/src/pages/About.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import Header from '../components/Header'
-
-function About() {
-  return (
-    <>
-      <Header />
-      <main className='main my-20'>
-          About
-      </main> 
-      </>
-  )
-}
-
-export default About
