+++ frontend/src/views/HomePage.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Footer } from '../components/footer/Footer'
+import { Header } from '../components/header/Header'
+
+export const HomePage = () => {
+  return (
+    <>
+        <Header />
+        <main className='main py-44'>Home</main>
+        <Footer />
+    </>
+  )
+}
