+++ frontend/src/views/LoginPage.js
@@ -1,12 +1,10 @@
 import React from 'react'
-import LoginForm from '../components/LoginForm'
+import { LoginForm } from '../components/login/LoginForm'
 
-function Login() {
+export const LoginPage = () => {
   return (
     <div className='w-full h-screen flex justify-center items-center sm:px-0 px-5'>
       <LoginForm />
     </div>
   )
 }
-
-export default Login
