+++ frontend/src/views/RegisterPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import {RegisterForm} from '../components/register/RegisterForm'
+export const RegisterPage = () => {
+  return (
+    <div className='w-full h-screen flex justify-center items-center sm:px-0 px-5'>
+        <RegisterForm/>
+    </div>
+  )
+}
